% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparqlchunks.R
\name{autoproxyconfig}
\alias{autoproxyconfig}
\title{Try to determine the proxy settings automatically}
\usage{
autoproxyconfig(endpoint)
}
\arguments{
\item{endpoint}{The SPARQL endpoint (URL)}
}
\description{
Try to determine the proxy settings automatically
}
