% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparqlchunks.R
\name{get_outcontent}
\alias{get_outcontent}
\title{Get the content from the endpoint}
\usage{
get_outcontent(endpoint, query, acceptype, proxy_config)
}
\arguments{
\item{endpoint}{The SPARQL endpoint (URL)}

\item{query}{The SPARQL query (character)}

\item{acceptype}{"text/csv" or "text/xml" (character)}

\item{proxy_config}{Detected proxy configuration (list)}
}
\description{
Get the content from the endpoint
}
