% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARQLchunks-package.R
\docType{package}
\name{SPARQLchunks-package}
\alias{SPARQLchunks}
\alias{SPARQLchunks-package}
\title{SPARQLchunks: Run 'SPARQL' Chunks and Inline Functions to Retrieve Data}
\description{
Allows to run 'SPARQL' chunks in R-markdown files. Also provides inline functions to send a 'SPARQL' query to an endpoint and retrieve data in data frame or list form.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aourednik/SPARQLchunks}
  \item Report bugs at \url{https://github.com/aourednik/SPARQLchunks/issues}
}

}
\author{
\strong{Maintainer}: Andre Ourednik \email{andre.ourednik@epfl.ch} (\href{https://orcid.org/0000-0002-0309-8715}{ORCID})

}
\keyword{internal}
