% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparqlchunks.R
\name{eng_sparql}
\alias{eng_sparql}
\title{Add \code{sparql} as knit-engine to knitr package}
\usage{
eng_sparql(options)
}
\arguments{
\item{opts}{Chunk options, as provided by \code{knitr} during chunk execution. Chunk option for this is \code{sparql}. Note that we avoid calling this "options" to avoid conflict with an R system function name.}
}
\value{
Data in dataframe or list form (depending on options)
}
\description{
Pointing knitr to this function alows you to run SPARQL chunks from R Markdown: \code{knitr::knit_engines$set(sparql = SPARQLchunks::eng_sparql)}. Usage is internal.
}
\details{
The \code{sparql} engine can be activated via

\if{html}{\out{<div class="sourceCode">}}\preformatted{knitr::knit_engines$set(sparql = SPARQLchunks::eng_sparql)
}\if{html}{\out{</div>}}

This will be set within an R Markdown document's setup chunk. Do not use eng_sparql function elsewhere.
}
\examples{
library(SPARQLchunks)
knitr::knit_engines$set(sparql = SPARQLchunks::eng_sparql)
}
\references{
André Ourednik (2021). Execute SPARQL chunks in Rmarkdown Available at:  https://ourednik.info/maps/2021/12/14/execute-sparql-chunks-in-r-markdown/
}
\author{
\href{https://ourednik.info}{André Ourednik}
}
\concept{important functions}
\keyword{documentation}
