% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparqlchunks.R
\name{sparql2df}
\alias{sparql2df}
\title{Fetch data from a SPARQL endpoint and store the output in a dataframe}
\usage{
sparql2df(endpoint, query, autoproxy = FALSE, auth = NULL)
}
\arguments{
\item{endpoint}{The SPARQL endpoint (a URL)}

\item{query}{The SPARQL query (character)}

\item{autoproxy}{Try to detect a proxy automatically (boolean). Useful on Windows machines behind corporate firewalls}

\item{auth}{Authentication Information (httr-authenticate-object)}
}
\value{
SPARQL query result in data.frame format
}
\description{
Fetch data from a SPARQL endpoint and store the output in a dataframe
}
\examples{
library(SPARQLchunks)
endpoint <- "https://lindas.admin.ch/query"
query <- "PREFIX schema: <http://schema.org/>
  SELECT * WHERE {
  ?sub a schema:DataCatalog .
  ?subtype a schema:DataType .
}"
result_df <- sparql2df(endpoint, query)
}
